Adapt inference and evaluation to whole volumes models  - Pad images to suitable spatial size w.r.t. model's filters kernels  - Inverse the padding after inference to retrieve original spatial size
N.B. : We have the option of a unique patch/non-patch behavior to suppress multiple tests on `is_patch`. Padding to a divisible value of patch-size and then reversing the transform after rebuilding the total volume (sliding_windowinferer) in case of a patch approach. Not fundamental work, but may be included in future updates.